requirements.txt.txt
python-telegram-bot==20.3
python-dotenv==1.0.1